Res 25

+++ src/server/router.c
@@ -1,10 +1,9 @@
 #include "router.h"
-#include "http.h"
 #include "http_errors.h"
 #include <string.h>
 
 // Given a set of routes, determine which route handler to use by comparing paths
-void find_route(Route* routes, int num_routes, HttpRequest* req, HttpResponse* resp) {
+void find_route(Route* routes, size_t num_routes, HttpRequest* req, HttpResponse* resp) {
     if (!req) {
         HttpResponse temp_resp = handle_404(req);
         *resp = temp_resp;
@@ -18,7 +17,7 @@ void find_route(Route* routes, int num_routes, HttpRequest* req, HttpResponse* r
         *query_start = '\0';
     }
 
-    for (int i = 0; i < num_routes; i++) {
+    for (size_t i = 0; i < num_routes; i++) {
         if (strcmp(clean_path, routes[i].path) == 0) {
             HttpResponse temp_resp = routes[i].handler(req);
             *resp = temp_resp;
