+++ src/utils/utils.c
@@ -14,8 +14,8 @@ char* skip_whitespace(char* src) {
 
 // REQUIRES: url encoded src string
 // EFFECTS: Returns the length of the decoded string
-int url_decoded_str_len(char* src) {
-    int len = 0;
+size_t url_decoded_str_len(char* src) {
+    size_t len = 0;
     while(*src) {
         if (*src == '%' && *(src + 1) && *(src + 2)) {
             len++;
@@ -31,9 +31,9 @@ int url_decoded_str_len(char* src) {
 // REQUIRES: Correctly allocated dst array, url encoded src string, length of src string
 // MODIFIES: Decodes src string into dst
 // EFFECTS: Decodes src string with url decode
-void decode_url(char* dst, char* src, int src_len) {
-    int dst_index = 0;
-    for (int src_index = 0; src_index < src_len; src_index++) {
+void decode_url(char* dst, char* src, size_t src_len) {
+    size_t dst_index = 0;
+    for (size_t src_index = 0; src_index < src_len; src_index++) {
         if (src[src_index] == '%' && src[src_index + 1] && src[src_index + 2]) {
             char hex[3] = {
                 src[src_index + 1], src[src_index + 2], '\0'
