+++ tests/testutils/src/test_http_utils.c
@@ -5,7 +5,7 @@
 void display_http_request(HttpRequest* req) {
     printf("--------------------------------------\n");
     printf("%s %s %s\n", req->method, req->path, req->version);
-    for (int i = 0; i < req->num_headers; i++) {
+    for (size_t i = 0; i < req->num_headers; i++) {
         HttpHeader cur_header = req->headers[i];
         printf("%s : %s\n", cur_header.key, cur_header.value);
     }
